currency in BigDecimal dates in LocalDate category to constant

+++ src/main/java/com/example/budgetkeeperspring/expense/MonthCategoryAmount.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -11,5 +13,5 @@ public class MonthCategoryAmount {
 
     private Integer month;
     private String category;
-    private Float amount;
+    private BigDecimal amount;
 }
